IGNITE-10968: [ML] Create new ignite module SparkMLModelImport and add LogRegression converter
This closes #5851
