+++ modules/ml/pom.xml
@@ -19,7 +19,8 @@
 <!--
     POM file.
 -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <properties>
         <commons.math.version>3.6.1</commons.math.version>
@@ -53,16 +54,16 @@
         </dependency>
 
         <dependency>
-	    <groupId>org.apache.ignite</groupId>
-	    <artifactId>ignite-indexing</artifactId>
-	    <version>${project.version}</version>
+            <groupId>org.apache.ignite</groupId>
+            <artifactId>ignite-indexing</artifactId>
+            <version>${project.version}</version>
         </dependency>
 
 
         <dependency>
-	    <groupId>org.apache.ignite</groupId>
-	    <artifactId>ignite-spring</artifactId>
-	    <version>${project.version}</version>
+            <groupId>org.apache.ignite</groupId>
+            <artifactId>ignite-spring</artifactId>
+            <version>${project.version}</version>
         </dependency>
 
         <dependency>
